แก้ไข
แก้formเพิ่มเครื่องจักร พิมพ์ไทยไม่ได้
แก้formแก้ไขเครื่องจักร พิมพ์ไทยไม่ได้
แก้formแก้ไขเครื่องจักร ข้อมูลไม่มาแสดง
แก้ลบเครื่องจักร ถามก่อน

+++ database/DeleteMachine.php
@@ -1,9 +1,21 @@
 <?php
 //ลบข้อมูลเครื่องจักร
     include 'connect.php';
-    $id = $_GET['id'];
+    if ( isset( $_POST['deletedata'] ) ) {
+    $id = $_POST['delete_id'];
     $query = "delete FROM tmstvehicle WHERE XVVehCode = $id;";
     $sql = mysqli_query($connect,$query);
     mysqli_close($connect);
-    echo "<script>window.location='../pages/ListMachine.php'</script>";
+    if ( $sql ) {
+        echo '<script>';
+        echo "alert('ทำการลบเครื่องจักรเรียบร้อยแล้ว !!!');";
+        echo "window.location='../pages/ListMachine.php';";
+        echo '</script>';
+    } else {
+        echo '<script>';
+        echo "alert('ไม่สามารถทำการลบเครื่องจักรได้ !!!');";
+         echo "window.location='../pages/ListMachine.php';";
+        echo '</script>';
+    }
+}
 ?>
